feat(kvbm): enable kvbm-v2 xpu-sycl
`lib/device/` (workspace crate `dynamo-device`) is the canonical
device abstraction for every kvbm-* crate.
  - `DeviceBackend::{Cuda, Sycl}` enum.
  - `Device{Context,Stream,Event,MemPool}Ops` traits in `traits.rs`,
    with `Device{Context,Stream,Event,MemPool}` trait-object wrapper
    structs (`Box<dyn ...Ops>`) re-exported from the crate root.
  - `DeviceEventOps::is_complete()` mirrors the oneapi-rs safe wrapper.
  - `DeviceStreamOps::wait_event(&dyn DeviceEventOps)` replaces direct
    `cuStreamWaitEvent` / `sycl_rs_queue_wait_event` calls.
  - `topology::get_device_cpu_set(backend, pci_address)` for
    multi-backend NUMA pinning.

Mutual exclusion: `cuda` and `xpu-sycl` features are guarded by
`compile_error!` in `lib/kvbm-connector/src/lib.rs`. Default features
in `lib/kvbm-connector/Cargo.toml` are `["cuda", "testing"]`; XPU is
opt-in via `--no-default-features --features xpu-sycl,testing`.

oneCCL collectives shim added alongside NCCL. `mod collectives` in
`lib/kvbm-engine/src/lib.rs` and `mod replicated` in
`lib/kvbm-engine/src/worker/physical.rs` are gated
`cfg(any(feature = "nccl", feature = "oneccl"))`; `cudarc` is now
`optional = true` in `kvbm-engine` and pulled only by `nccl`.

`lib/kvbm-connector/src/connector/worker/{mod,state}.rs` stores
`Arc<DeviceEvent>` / `Arc<DeviceStream>`; the
`trigger_forward_pass_completion` poll loop uses
`DeviceEvent::is_complete()`, preserving `cuEventQuery` shape poll.

`lib/kvbm-engine/src/worker/physical.rs::execute_local_layerwise_onboard`
is the multi-backend form: `&[Arc<DeviceEvent>]` +
`acquire_h2d_stream()` + `event.record_on(&stream)`.

`lib/kvbm-engine/src/collectives/bootstrap.rs` -- explicit
`as u8` cast added so the byte serialiser builds on architectures
where `c_char` defaults to `u8` (aarch64) as well as `i8` (x86_64).
Was masked previously by the umbrella `collectives = ["nccl"]` only
ever pairing with x86_64 `cuda`.

`CudaAsync{H2D,D2H,D2D}` -> `Async{H2D,D2H,D2D}` and
`CudaBlocking{H2D,D2H}` -> `Blocking{H2D,D2H}` in
`lib/kvbm-physical/src/transfer/`. Variants are backend-agnostic in
semantics, multi-backend executor dispatches via `DeviceBackend`.

`(System, Device)` and `(Device, System)` strategies in
`transfer/strategy.rs` previously panicked with "not supported"; they
now lower to `Direct(BlockingH2D)` / `Direct(BlockingD2H)`. The
local-transfer test fixture (`build_agent_for_kinds` in
`lib/kvbm-physical/src/transfer/tests/local_transfers.rs`) no longer
auto-registers UCX for `StorageKind::Device(_)`: local H2D / D2H / D2D
go through native SYCL / CUDA APIs and don't need a NIXL backend at all.
UCX is still used for remote RDMA paths on both CUDA and XPU hosts.

SYCL permute / vectorized-copy kernels added at
`lib/kvbm-kernels/sycl/{tensor_permute_kernel,vectorized_copy_kernel}.cpp`,
Rust glue in `tensor_kernels_sycl.rs`. CUDA permute kernels in
`cuda/tensor_kernels.cu` are unchanged.

Host-staged d2dx path in `examples/kvbench_xpu_sycl.rs`.

`bench_engine` / `bench_collectives` / `bench_transfer` / `kvbench`
print selected-device summaries at startup. `bench_engine` is the
only consumer of the new backend-aware
`dynamo_device::topology::get_device_cpu_set(backend, pci_address)`
helper today (one call site at `lib/kvbm-engine/bin/bench_engine.rs:679`
for PCI-BDF-based NUMA pinning); the other three benchmarks pin via
existing routes. `cpu_slices_by_pci()` stays public for the
backend-aware overload to call back into.

`validate_numa_placement` moved from `lib/memory/bin/` to
`lib/kvbm-physical/bin/` and extended for both backends.

`lib/memory/src/tests.rs` -- `mod cuda_tests` (gated `testing-cuda`)
renamed to `mod device_tests` and `mod cuda_nixl_tests` (gated
`testing-all`, nested under `mod nixl_tests`) renamed to
`mod device_nixl_tests`; both regated
`cfg(any(feature = "testing-cuda", feature = "testing-xpu-sycl"))` to
match `test_device_ctx()`'s definition surface (so `--features
testing-nixl` alone still compiles).

DIS-1310 -- `NixlAgent::add_backend_with_params` now merges
`custom_params` into plugin defaults via `nixl_sys 0.10+` `Params::clone`
+ `Params::set` instead of bailing with "not yet supported". Tests
inverted (`_fails` -> success). Behaviour change not XPU-driven.

`lib/memory/src/numa/worker_pool.rs` -- restored XPU's
`pub trait PinnedAllocator { alloc_pinned, free_pinned }`. The helper
itself is renamed `allocate_pinned_for_gpu` -> `allocate_pinned_for_device`.
Call sites in `lib/device/src/{cuda,sycl}/mod.rs`, plus the worker_pool
tests and the `kvbm_v2_xpu_sycl_enablement.md` class diagram updated.

`lib/memory/src/numa/mod.rs` -- single sysfs walk
(`cpu_slices_by_pci()`); vendor filter widened to include Intel
`0x8086` so SYCL GPUs get NUMA-pinned the same way CUDA GPUs do.
Resolved a duplicate `get_pci_bus_address_from_cuda` and a missing
`cudarc::driver::result::device` import inside `enumerate_cuda_gpus`.

`lib/device/src/topology.rs` -- collapsed to a wrapper that
reads from engine-service's PCI cache:
`cpu_slices_by_pci().get(pci_address).cloned()`. The XPU side's
per-backend enumerator dispatch is no longer needed, the unified
sysfs walk covers both vendors.

`lib/memory/src/lib.rs` -- restored `StorageError::Cuda(#[from]
cudarc::driver::DriverError)` so engine-service's CUDA-only
`mmap_pinned` keeps compiling.

`lib/memory/src/device.rs` -- restored `pub(crate) fn
cuda_context(device_id)` helper for engine-service's CUDA-only
`mmap_pinned`. Cached `OnceLock<Mutex<HashMap<u32, Arc<CudaContext>>>>`
matches the existing module style.

`lib/memory/Cargo.toml` -- the engine-service `[[bin]]
validate_numa_placement` entry was remove. The underlying source
file `lib/memory/bin/validate_numa_placement.rs` was deleted because
the engine-service version depends on the CUDA-only
`PinnedStorage::new_for_device(len, Some(gpu_id))` API which doesn't
exist in the XPU allocator-based shape. The backend-agnostic restoration
lives at `lib/kvbm-physical/bin/validate_numa_placement.rs` and is registered
under that crate.

+++ lib/kvbm-engine/src/collectives/bootstrap.rs
@@ -131,7 +131,7 @@ impl NcclBootstrap {
         let mut bytes = Vec::with_capacity(8 + 128);
         bytes.extend_from_slice(&(self.world_size as u64).to_le_bytes());
         for &byte in &self.nccl_id.internal {
-            bytes.push(byte);
+            bytes.push(byte as u8);
         }
         bytes
     }
